requirements.txt
asgiref==3.5.2
Django==4.1.1
django-ckeditor==6.5.0
django-js-asset==2.0.0
django-ninja==0.19.1
Pillow==9.2.0
pydantic==1.10.2
sqlparse==0.4.2
typing_extensions==4.3.0
tzdata==2022.2
python-jose==3.3.0